+++ src/SignatureValidator/DefaultSignatureValidator.php
@@ -3,6 +3,7 @@
 namespace Spatie\WebhookClient\SignatureValidator;
 
 use Illuminate\Http\Request;
+use Spatie\WebhookClient\Events\InvalidSignatureEvent;
 use Spatie\WebhookClient\WebhookConfig;
 use Spatie\WebhookClient\Exceptions\WebhookFailed;
 
@@ -11,6 +12,10 @@ class DefaultSignatureValidator implements SignatureValidator
     public function isValid(Request $request, WebhookConfig $config): bool
     {
         $signature = $request->header($config->signatureHeaderName);
+
+        if (! $signature) {
+           return false;
+        }
 
         $signingSecret = $config->signingSecret;