+++ tests/TestClasses/NothingIsValidSignatureValidator.php
@@ -1,0 +1,17 @@
+<?php
+
+
+namespace Spatie\WebhookClient\Tests\TestClasses;
+
+
+use Illuminate\Http\Request;
+use Spatie\WebhookClient\SignatureValidator\SignatureValidator;
+use Spatie\WebhookClient\WebhookConfig;
+
+class NothingIsValidSignatureValidator implements SignatureValidator
+{
+    public function isValid(Request $request, WebhookConfig $config): bool
+    {
+        return false;
+    }
+}